# internal/entity/team.go
package entity

type TeamMember struct {
	UserID   	string `json:"user_id"`
	Username 	string `json:"username"`
	IsActive 	bool   `json:"is_active"`
}

type Team struct {
	TeamName 	string       `json:"team_name"`
	Members  	[]TeamMember `json:"members"`
}